requirements.txt
django_filter==2.4.0
Django==3.0.7
djangorestframework==3.12.2
Pillow==8.0.1
django-crispy-forms==1.9.2
